New variable engine implemented Few words now you can set variable instant in one char without create new column in database
For Example setVar("isNoob", "true") that means it will save variable for this player as noob and type will be boolean

and you can check it by using player.getVarB("isNoob", "false")
it will return is character has isNoob value if is true will return true and vice versa, if this var is null will return false value by default
if you make it player.getVarB("isNoob", "true") and var isnoob is empty or null will return true.

Implemented new xml parasers and holders

For those engine we will create documentation in forum later

Core changes:
compile and copy paste core.jar and engine.jar

Data changes:
Just execute new sql file sunrise_variables.sql located in tools/sql/sunrise_sql_files/

+++ L2J_SunriseProject_Core/java/gr/sr/main/Conditions.java
@@ -209,7 +209,7 @@ public class Conditions
 	 * @param count
 	 * @return boolean
 	 */
-	public static boolean checkPlayerItemCount(L2PcInstance player, int itemId, int count)
+	public static boolean checkPlayerItemCount(L2PcInstance player, int itemId, long count)
 	{
 		if ((player.getInventory().getItemByItemId(itemId) == null) || (player.getInventory().getItemByItemId(itemId).getCount() < count))
 		{
